+++ Runtime/Scripts/TetraGenDataTypes.cs
@@ -9,8 +9,9 @@ namespace TetraGen
     [System.Serializable]
     public class ChunkData
     {
-        public List<GameObject> meshes = new List<GameObject>();
+        public readonly List<GameObject> meshes = new List<GameObject>();
     }
+
     /// <summary> Representation of Signed Distance shapes for GPU processing. </summary>
     [StructLayout(LayoutKind.Sequential)]
     public struct ShapeData